fix: check if port 80 or 443 is already in use before starting Envoy
Envoy doesn't exit if it can't bind ports for all its listeners, and I couldn't find any configuration option to make it do so.

With this, auto-restart will kick in until both ports are no longer in use.

May not be necessary after the previous fix to ensure that Kong service is stopped before Envoy service is started, but doesn't hurt to have multiple defensive measures, since it's somewhat unexpected that Envoy doesn't exit on such failures.