fix: ensure that we pull apt updates before building
I'm not quite certain why the current apt commands are structured the
way they are, pulling the package lists _after_ operations that use
the package list.

The proposed change should unblock the build, but we should clean this
up during the week once folks that contributed the original code are
available.