chore(qemu): ship the initialized data dir at a different location
The reason for not shipping it originally was to avoid any possibility
of the server starting up off a datadir on ephemeral storage.

The initdb process itself is decently heavy though. We can tread the
middle ground between the two by shipping the datadir at a secondary
location, which the server never gets configured to run off. This
datadir can be moved to the appropriate place on persistent storage
prior to server startup. The simpler operation of copying the files
over should be faster than the initdb routine.