docs: generate documentation site with MkDocs
Add MkDocs configuration and assets for generating a documentation site.
This is more readable and provides a structured way to present information about the project.

We don't store the generated site in the repository, but instead use MkDocs to build it on demand.

In a development shell, one can run `mkdocs serve` to quickly preview the site locally.