feat: disable default database collector for postgres_exporter
The database collector emits a metric for the size of each database by
default. We emit a metric that represents the sum of all database,
which suffices for our purposes. This change removes the redundant
per-database metric.

In order to avoid a change in the AMI metrics, the change is only made
for the QEMU artifact.