feat(pg_net): switch version on file system using overlayfs
For pg_net background worker, we need to switch the version of the
extension on the file system (both pg_net.so and pg_net.control need to point
to the desired version).

As the extension is in the Nix store, we cannot simply symlink to the
desired version, as the Nix store is read-only. 
To work around this, we use overlayfs to create a writable layer on top
of the pg_net store path.