fix: reduce pressure on CI runners
Updating the dependency helping us build all our rust extensions has put
a lot of pressure on our CI runners, causing timeouts and failed builds.

We are currently working on a more permanent solution in #1745 but in
the meantime, we cap the number of parallel nix builds to 2, which
should reduce the load on our CI runners.