refactor: migrate from packages to legacyPackages for PostgreSQL extensions
To be able to build extensions versions packages separately in CI, we
need to expose them in a nested structure. It is currently not possible
to do so with the flattened packages structure, as the individual
extension packages are not directly accessible.

In this change, we replace the flattened package structure with nested
legacyPackages to improve discoverability of individual extension
packages.