refactor(devshell): improve devshell experience
The current development shell include a large number of packages which
leads to a very large closure size (5 Gb) and slow startup times.
To improve the situation we defer installing large packages to devshell
commands which only get installed when needed.
We introduce the use of devshell to define the commands in a
declarative way as well as to improve the overall developer experience.

Included changes:

- Add organized command menu with categories (check, ami, extension, postgres)
- Add watch command using watchexec + nix-fast-build for continuous checking
- Expose common tasks as named commands: fmt, check, lint, watch
- Add aws-vault to development tools